fix: bedrock updates
