+++ cradle/windows/Cargo.template.toml
@@ -13,7 +13,6 @@ debug = ["peridot/debug"]
 mt = ["peridot/mt"]
 
 [dependencies]
-bedrock = { git = "https://github.com/Pctg-x8/bedrock", branch = "peridot", features = ["Presentation", "Implements", "VK_KHR_win32_surface"] }
 peridot = { path = "../.." }
 log = "0.4"
 env_logger = "0.5"
@@ -23,6 +22,22 @@ comdrive = { git = "https://github.com/Pctg-x8/comdrive", optional = true }
 metrics = { git = "https://github.com/Pctg-x8/metrics", optional = true }
 widestring = { version = "0.4", optional = true }
 parking_lot = "0.11"
+
+[dependencies.bedrock]
+git = "https://github.com/Pctg-x8/bedrock"
+branch = "peridot"
+features = [
+    "Presentation",
+    "Implements",
+    "VK_KHR_win32_surface",
+    "VK_KHR_display",
+    "VK_EXT_full_screen_exclusive",
+    "VK_KHR_get_physical_device_properties2",
+    "VK_KHR_get_surface_capabilities2"
+]
+
+[build-dependencies]
+winres = "0.1.12"
 
 [dependencies.winapi]
 version = "0.3"
