+++ cradle/windows/app.rc
@@ -1,0 +1,11 @@
+#include <Windows.h>
+
+1001 DIALOGEX 0, 0, 310, 47
+STYLE DS_SETFONT | DS_MODALFRAME | DS_FIXEDSYS | DS_CENTER | WS_POPUP | WS_CAPTION | WS_SYSMENU
+CAPTION "Select Monitor"
+FONT 8, "MS Shell Dlg", 400, 0, 0x1
+BEGIN
+    DEFPUSHBUTTON   "&Launch",1,103,26,50,14
+    PUSHBUTTON      "&Exit",2,157,26,50,14
+    COMBOBOX        10,7,7,296,30,CBS_DROPDOWNLIST | CBS_HASSTRINGS | WS_VSCROLL | WS_TABSTOP
+END
